Refactor table-export
