+++ ui-frontend/packages/catalog-ui-search/src/main/webapp/component/result-selector/paging.tsx
@@ -3,11 +3,8 @@ import { hot } from 'react-hot-loader'
 import Button from '@material-ui/core/Button'
 import { useBackbone } from '../selection-checkbox/useBackbone.hook'
 import { useLazyResultsStatusFromSelectionInterface } from '../selection-interface/hooks'
-import { Elevations } from '../theme/theme'
-import CloseIcon from '@material-ui/icons/Close'
 import TableExport from '../table-export/table-export'
 import { useDialog } from '../dialog'
-import { DarkDivider } from '../dark-divider/dark-divider'
 
 type Props = {
   selectionInterface: any
@@ -85,46 +82,13 @@ const Paging = ({ selectionInterface }: Props) => {
         disabled={isSearching}
         onClick={() => {
           dialogContext.setProps({
-            PaperProps: {
-              style: {
-                minWidth: 'none',
-              },
-              elevation: Elevations.panels,
-            },
-            open: true,
-            disableEnforceFocus: true, // otherwise we can't click inside 3rd party libraries using portals (like date picker from blueprint)
             children: (
-              <div
-                className="min-w-screen-1/2"
-                style={{
-                  minHeight: '60vh',
-                }}
-              >
-                <div className="text-2xl text-center px-2 pb-2 pt-4 font-normal relative">
-                  Export
-                  <Button
-                    data-id="close-button"
-                    className="absolute right-0 top-0 mr-1 mt-1"
-                    variant="text"
-                    color="default"
-                    size="small"
-                    onClick={() => {
-                      dialogContext.setProps({
-                        open: false,
-                        children: null,
-                      })
-                    }}
-                  >
-                    <CloseIcon />
-                  </Button>
-                </div>
-                <DarkDivider className="w-full h-min" />
-                <TableExport
-                  selectionInterface={selectionInterface}
-                  filteredAttributes={[]}
-                />
-              </div>
+              <TableExport
+                selectionInterface={selectionInterface}
+                filteredAttributes={[]}
+              />
             ),
+            open: true,
           })
         }}
         color="primary"
