+++ ui-frontend/packages/catalog-ui-search/src/main/webapp/react-component/alert-settings/presentation.tsx
@@ -97,9 +97,7 @@ const render = (props: Props) => {
               onPersistenceChange(newValue.value)
             }}
             getOptionSelected={(option) => option.value === persistence}
-            getOptionLabel={(option) => {
-              return option.label
-            }}
+            getOptionLabel={(option) => option.label}
             disableClearable
             value={keepNotificationsOptions.find(
               (choice) => choice.value === persistence
@@ -122,9 +120,7 @@ const render = (props: Props) => {
                 onExpirationChange(newValue.value)
               }}
               getOptionSelected={(option) => option.value === expiration}
-              getOptionLabel={(option) => {
-                return option.label
-              }}
+              getOptionLabel={(option) => option.label}
               disableClearable
               value={expireOptions.find(
                 (choice) => choice.value === expiration
