+++ ui-frontend/packages/catalog-ui-search/src/main/webapp/react-component/utils/export/index.tsx
@@ -21,4 +21,8 @@ export {
   ResultSet,
   ExportCountInfo,
   DownloadInfo,
+  Option,
+  ExportResponse,
+  getWarning,
+  generateOnDownloadClick,
 } from './export'
